Added image function to posts and some other minor fixes

+++ app/Http/Controllers/PostsController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
+use Auth;
 
 class PostsController extends Controller
 {
@@ -12,6 +14,12 @@ class PostsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         // $posts = Post::all();
@@ -39,13 +47,34 @@ class PostsController extends Controller
     {
         $this->validate($request, [
                 'title' => 'required',
-                'body' => 'required'
+                'body' => 'required',
+                'cover_image' => 'image|nullable|max:1999'
         ]);
+
+        // Handle file upload
+        if($request->hasFile('cover_image')) {
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+            // Get just ext
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            //Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            //Upload the image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         //Create a post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
+        $post->user_id = Auth::user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success', 'Post Created');
@@ -72,6 +101,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -89,10 +124,29 @@ class PostsController extends Controller
             'body' => 'required'
     ]);
 
+    // Handle file upload
+    if($request->hasFile('cover_image')) {
+        // Get filename with the extension
+        $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+        
+        // Get just filename
+        $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+
+        // Get just ext
+        $extension = $request->file('cover_image')->getClientOriginalExtension();
+        //Filename to store
+        $fileNameToStore = $filename.'_'.time().'.'.$extension;
+        //Upload the image
+        $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+    }
+
     //Create a post
     $post = Post::find($id);
     $post->title = $request->input('title');
     $post->body = $request->input('body');
+    if($request->hasFile('cover_image')) {
+        $post->cover_image = $fileNameToStore;
+    }
     $post->save();
 
     return redirect('/posts')->with('success', 'Post Updated');
@@ -107,8 +161,18 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        
+        //Check for correct user
+        if(auth()->user()->id !==$post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized page');
+        }
+
+        if($post->cover_image != 'noimage.jpg') {
+            // Delete the image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
 
-        return redirect('/posts')->with('success', 'Post Deleted    ');
+        return redirect('/posts')->with('success', 'Post Deleted');
     }
 }
